// wordle/src/features/wordle/types.d.ts
export type Char = 'a' | 'b' | 'c' | 'd' | 'e' | 'f' | 'g' | 'h' | 'i' | 'j' | 'k' 
| 'l' | 'm' | 'n' | 'o' | 'p' | 'q' | 'r' | 's' | 't' | 'u' | 'v' | 'w' | 'x' 
| 'y' | 'z' | 'A' | 'B' | 'C' | 'D' | 'E' | 'F' | 'G' | 'H' | 'I' | 'J' | 'K' 
| 'L' | 'M' | 'N' | 'O' | 'P' | 'Q' | 'R' | 'S' | 'T' | 'U' | 'V' | 'W' | 'X' 
| 'Y' | 'Z' | '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '' 


export interface Board {
  lines: Line[5]; 
}

interface Sorted {
  position: number;
}
export interface Line extends Sorted{
  boxes: Box[5];
}

export interface Box extends Sorted{
  letter: Char;
  filled: boolean;
  fullMatch: boolean;
  match: boolean;
}

type index = 1 | 2 | 3 | 4 | 5;

export interface Position {
  lineIdx: index;
  boxIdx: index;
}